+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/rntbd/RntbdConstants.java
@@ -598,8 +598,8 @@ public final class RntbdConstants {
         PriorityLevel((short) 0x00BF, RntbdTokenType.Byte, false),
         GlobalDatabaseAccountName((short) 0x00CE, RntbdTokenType.String, false),
         ThroughputBucket((short)0x00DB, RntbdTokenType.Byte, false),
-        HubRegionProcessingOnly((short)0x00EF, RntbdTokenType.Byte , false),
-        WorkloadId((short)0x00DC, RntbdTokenType.Byte, false);
+        WorkloadId((short)0x00DC, RntbdTokenType.Byte, false),
+        HubRegionProcessingOnly((short)0x00EF, RntbdTokenType.Byte , false);
 
         public static final List<RntbdRequestHeader> thinClientHeadersInOrderList = Arrays.asList(
             EffectivePartitionKey,